requirements.txt
streamlit
google-generativeai
gspread
oauth2client
Pillow